+++ python/dna_semantic_compression/__init__.py
@@ -1,1 +1,0 @@
-from .dna_semantic_compression import *  # noqa: F403